# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NexFlow CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: segment_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactListResponse"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactBody"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /contacts/{id}/activities:
    get:
      operationId: getContactActivities
      tags: [contacts]
      summary: Get contact activity timeline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"

  # ── Companies ──
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List companies
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyListResponse"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyBody"
      responses:
        "201":
          description: Created company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompanyBody"
      responses:
        "200":
          description: Updated company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    delete:
      operationId: deleteCompany
      tags: [companies]
      summary: Delete company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Deals / Pipeline ──
...
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List activities
      parameters:
        - name: contact_id
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Create activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"

  # ── Signals ──
...
  /calls:
    get:
      operationId: listCalls
      tags: [calls]
      summary: List calls
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of calls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Call"
    post:
      operationId: createCall
      tags: [calls]
      summary: Initiate call
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCallBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
  /calls/{id}:
    get:
      operationId: getCall
      tags: [calls]
      summary: Get call
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"

  # ── Scripts ──
...
components:
  schemas:
...
    Contact:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        phone:
          type: string
        title:
          type: string
        company_id:
          type: string
        company_name:
          type: string
        lead_score:
          type: number
        status:
          type: string
          enum: [new, active, qualified, unqualified, customer]
        avatar_url:
          type: string
        linkedin_url:
          type: string
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - first_name
        - last_name
        - status
        - created_at
    ContactListResponse:
      type: object
      properties:
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        total:
          type: integer
      required:
        - contacts
        - total
    CreateContactBody:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        phone:
          type: string
        title:
          type: string
        company_id:
          type: string
        status:
          type: string
          enum: [new, active, qualified, unqualified, customer]
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - first_name
        - last_name
    UpdateContactBody:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        phone:
          type: string
        title:
          type: string
        company_id:
          type: string
        status:
          type: string
          enum: [new, active, qualified, unqualified, customer]
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
        lead_score:
          type: number
    Company:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        name:
          type: string
        domain:
          type: string
        industry:
          type: string
        size:
          type: string
          enum: [1-10, 11-50, 51-200, 201-500, 501-1000, 1000+]
        country:
          type: string
        revenue:
          type: integer
          description: Revenue in smallest currency unit
        logo_url:
          type: string
        linkedin_url:
          type: string
        website:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        contact_count:
          type: integer
        deal_count:
          type: integer
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - name
        - created_at
    CompanyListResponse:
      type: object
      properties:
        companies:
          type: array
          items:
            $ref: "#/components/schemas/Company"
        total:
          type: integer
      required:
        - companies
        - total
    CreateCompanyBody:
      type: object
      properties:
        name:
          type: string
        domain:
          type: string
        industry:
          type: string
        size:
          type: string
          enum: [1-10, 11-50, 51-200, 201-500, 501-1000, 1000+]
        country:
          type: string
        revenue:
          type: integer
        website:
          type: string
        description:
          type: string
      required:
        - name
    UpdateCompanyBody:
      type: object
      properties:
        name:
          type: string
        domain:
          type: string
        industry:
          type: string
        size:
          type: string
          enum: [1-10, 11-50, 51-200, 201-500, 501-1000, 1000+]
        country:
          type: string
        revenue:
          type: integer
        website:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
...
    Activity:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        contact_id:
          type: string
        contact_name:
          type: string
        deal_id:
          type: string
        type:
          type: string
          enum: [call, email, whatsapp, meeting, note, task]
        title:
          type: string
        body:
          type: string
        status:
          type: string
          enum: [pending, completed, cancelled]
        scheduled_at:
          type: string
        completed_at:
          type: string
        created_at:
          type: string
      required:
        - id
        - type
        - title
        - status
        - created_at
    CreateActivityBody:
      type: object
      properties:
        contact_id:
          type: string
        deal_id:
          type: string
        type:
          type: string
          enum: [call, email, whatsapp, meeting, note, task]
        title:
          type: string
        body:
          type: string
        status:
          type: string
          enum: [pending, completed, cancelled]
        scheduled_at:
          type: string
      required:
        - type
        - title
...
    Call:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        contact_id:
          type: string
        contact_name:
          type: string
        direction:
          type: string
          enum: [inbound, outbound]
        status:
          type: string
          enum: [scheduled, in_progress, completed, missed, failed]
        duration_seconds:
          type: integer
        recording_url:
          type: string
        transcript:
          type: string
        sentiment_score:
          type: number
        call_score:
          type: number
        coaching_notes:
          type: string
        started_at:
          type: string
        ended_at:
          type: string
        created_at:
          type: string
      required:
        - id
        - status
        - created_at
    CreateCallBody:
      type: object
      properties:
        contact_id:
          type: string
        direction:
          type: string
          enum: [inbound, outbound]
        scheduled_at:
          type: string
      required:
        - contact_id
...