# lib/api-zod/src/generated/types/createActivityBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NexFlow CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateActivityBodyStatus } from "./createActivityBodyStatus";
import type { CreateActivityBodyType } from "./createActivityBodyType";

export interface CreateActivityBody {
  contact_id?: string;
  deal_id?: string;
  type: CreateActivityBodyType;
  title: string;
  body?: string;
  status?: CreateActivityBodyStatus;
  scheduled_at?: string;
}