# lib/api-zod/src/generated/types/createCompanyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NexFlow CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCompanyBodySize } from "./createCompanyBodySize";

export interface CreateCompanyBody {
  name: string;
  domain?: string;
  industry?: string;
  size?: CreateCompanyBodySize;
  country?: string;
  revenue?: number;
  website?: string;
  description?: string;
}